Rename llm response classes
