+++ lib/langchain.rb
@@ -16,10 +16,12 @@ loader.inflector.inflect(
   "llm" => "LLM",
   "openai" => "OpenAI",
   "openai_validator" => "OpenAIValidator",
+  "openai_response" => "OpenAIResponse",
   "pdf" => "PDF",
   "react_agent" => "ReActAgent",
   "sql_query_agent" => "SQLQueryAgent"
 )
+loader.collapse("#{__dir__}/langchain/llm/response")
 loader.setup
 
 # Langchain.rb a is library for building LLM-backed Ruby applications. It is an abstraction layer that sits on top of the emerging AI-related tools that makes it easy for developers to consume and string those services together.
