+++ lib/langchain/llm/google_palm.rb
@@ -34,12 +34,12 @@ module Langchain::LLM
     # Generate an embedding for a given text
     #
     # @param text [String] The text to generate an embedding for
-    # @return [Langchain::LLM::Response::GooglePalm] LLM response object with embedding as a value
+    # @return [Langchain::LLM::GooglePalmResponse] LLM response object with embedding as a value
     #
     def embed(text:)
       response = client.embed(text: text)
 
-      Langchain::LLM::Response::GooglePalm.new response, model: @defaults[:embeddings_model_name], type: "embedding"
+      Langchain::LLM::GooglePalmResponse.new response, model: @defaults[:embeddings_model_name], type: "embedding"
     end
 
     #
@@ -47,7 +47,7 @@ module Langchain::LLM
     #
     # @param prompt [String] The prompt to generate a completion for
     # @param params extra parameters passed to GooglePalmAPI::Client#generate_text
-    # @return [Langchain::LLM::Response::GooglePalm] LLM response object with completion as a value
+    # @return [Langchain::LLM::GooglePalmResponse] LLM response object with completion as a value
     #
     def complete(prompt:, **params)
       default_params = {
@@ -68,7 +68,7 @@ module Langchain::LLM
 
       response = client.generate_text(**default_params)
 
-      Langchain::LLM::Response::GooglePalm.new response, model: default_params[:model], type: "completion"
+      Langchain::LLM::GooglePalmResponse.new response, model: default_params[:model], type: "completion"
     end
 
     #
@@ -79,7 +79,7 @@ module Langchain::LLM
     # @param context [String] An initial context to provide as a system message, ie "You are RubyGPT, a helpful chat bot for helping people learn Ruby"
     # @param examples [Array<Hash>] Examples of messages to provide to the model. Useful for Few-Shot Prompting
     # @param options [Hash] extra parameters passed to GooglePalmAPI::Client#generate_chat_message
-    # @return [Langchain::LLM::Response::GooglePalm] LLM response object with response messages as a value
+    # @return [Langchain::LLM::GooglePalmResponse] LLM response object with response messages as a value
     #
     def chat(prompt: "", messages: [], context: "", examples: [], **options)
       raise ArgumentError.new(":prompt or :messages argument is expected") if prompt.empty? && messages.empty?
@@ -108,7 +108,7 @@ module Langchain::LLM
       response = client.generate_chat_message(**default_params)
       raise "GooglePalm API returned an error: #{response}" if response.dig("error")
 
-      Langchain::LLM::Response::GooglePalm.new response, model: default_params[:model], type: "chat.completion"
+      Langchain::LLM::GooglePalmResponse.new response, model: default_params[:model], type: "chat.completion"
     end
 
     #
